requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.3
django-ckeditor==6.3.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
idna==3.4
Pillow==9.3.0
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
